Modifications regarding remediation audit type and "ready for retest" functionality
I added a "ready for retest" toggle next to the "completed" toggle so a user can mark a finding as ready for retest or not. I also added a check icon on the side bar that works with the toggle for it to be more visual.

Also, when the audit type is "remediation", I changed the edit UI as all these fields are not necessaryfor a remediation report.